Optimize scene 7 tree-shaking
Split GLB parsing and single-light WGSL helpers so .gltf and hemispheric-light scenes only load needed chunks.
