+++ packages/babylon-lite/src/loader-gltf/gltf-parser.ts
@@ -1,6 +1,5 @@
 /**
  * Low-level glTF/GLB parsing helpers:
- * - GLB binary container parsing
  * - Accessor resolution (buffer views → typed arrays)
  * - Image extraction (embedded or external)
  * - Node hierarchy traversal with memoized world-matrix computation
@@ -25,41 +24,6 @@ const TYPE_SIZES: Record<string, number> = {
     MAT3: 9,
     MAT4: 16,
 };
-
-// --- GLB Container Parsing ---
-
-export function parseGlbContainer(buffer: ArrayBuffer): { json: any; binChunk: DataView } {
-    const view = new DataView(buffer);
-
-    // Header (12 bytes)
-    const magic = view.getUint32(0, true);
-    if (magic !== 0x46546c67) {
-        throw new Error("Not a valid GLB file");
-    }
-    // const version = view.getUint32(4, true);
-    // const totalLength = view.getUint32(8, true);
-
-    // JSON chunk
-    let offset = 12;
-    const jsonLength = view.getUint32(offset, true);
-    const jsonType = view.getUint32(offset + 4, true);
-    if (jsonType !== 0x4e4f534a) {
-        throw new Error("First GLB chunk is not JSON");
-    }
-    const jsonStr = new TextDecoder().decode(new Uint8Array(buffer, offset + 8, jsonLength));
-    const json = JSON.parse(jsonStr);
-    offset += 8 + jsonLength;
-
-    // BIN chunk
-    const binLength = view.getUint32(offset, true);
-    const binType = view.getUint32(offset + 4, true);
-    if (binType !== 0x004e4942) {
-        throw new Error("Second GLB chunk is not BIN");
-    }
-    const binChunk = new DataView(buffer, offset + 8, binLength);
-
-    return { json, binChunk };
-}
 
 // --- Accessor Resolution ---
